Rename 'serialization.py' to 'serializer.py'
