+++ django_modern_rest/metadata.py
@@ -23,7 +23,7 @@ if TYPE_CHECKING:
     from django_modern_rest.parsers import Parser
     from django_modern_rest.renderers import Renderer
     from django_modern_rest.security.base import AsyncAuth, SyncAuth
-    from django_modern_rest.serialization import BaseSerializer
+    from django_modern_rest.serializer import BaseSerializer
     from django_modern_rest.settings import HttpSpec
 
 ComponentParserSpec: TypeAlias = tuple[type['ComponentParser'], tuple[Any, ...]]
