+++ django_modern_rest/negotiation.py
@@ -16,7 +16,7 @@ from django_modern_rest.internal.negotiation import (
 from django_modern_rest.metadata import EndpointMetadata
 from django_modern_rest.parsers import Parser
 from django_modern_rest.renderers import Renderer
-from django_modern_rest.serialization import BaseSerializer
+from django_modern_rest.serializer import BaseSerializer
 
 
 class RequestNegotiator:
