+++ django_modern_rest/options_mixins.py
@@ -9,7 +9,7 @@ from django_modern_rest.metadata import ResponseSpec
 
 if TYPE_CHECKING:
     from django_modern_rest.controller import Controller
-    from django_modern_rest.serialization import BaseSerializer
+    from django_modern_rest.serializer import BaseSerializer
 
 
 #: Metadata for the default options response.
