+++ django_modern_rest/parsers.py
@@ -12,7 +12,7 @@ from django_modern_rest.exceptions import (
 from django_modern_rest.metadata import ResponseSpec, ResponseSpecProvider
 
 if TYPE_CHECKING:
-    from django_modern_rest.serialization import BaseSerializer
+    from django_modern_rest.serializer import BaseSerializer
 
 #: Types that are possible to load json from.
 Raw: TypeAlias = str | bytes | bytearray
