+++ django_modern_rest/plugins/msgspec/serializer.py
@@ -11,7 +11,7 @@ from typing_extensions import TypedDict, override
 
 from django_modern_rest.parsers import Parser, Raw
 from django_modern_rest.renderers import Renderer
-from django_modern_rest.serialization import (
+from django_modern_rest.serializer import (
     BaseEndpointOptimizer,
     BaseSerializer,
 )
