+++ django_modern_rest/renderers.py
@@ -11,7 +11,7 @@ from django_modern_rest.exceptions import NotAcceptableError
 from django_modern_rest.metadata import ResponseSpec, ResponseSpecProvider
 
 if TYPE_CHECKING:
-    from django_modern_rest.serialization import BaseSerializer
+    from django_modern_rest.serializer import BaseSerializer
 
 
 class Renderer(ResponseSpecProvider):
