+++ django_modern_rest/security/http.py
@@ -15,7 +15,7 @@ from django_modern_rest.security.base import AsyncAuth, SyncAuth
 if TYPE_CHECKING:
     from django_modern_rest.controller import Controller
     from django_modern_rest.endpoint import Endpoint
-    from django_modern_rest.serialization import BaseSerializer
+    from django_modern_rest.serializer import BaseSerializer
 
 
 class _HttpBasicAuth:
