+++ django_modern_rest/validation/controller.py
@@ -2,7 +2,7 @@ from collections.abc import Set
 from typing import TYPE_CHECKING
 
 from django_modern_rest.exceptions import EndpointMetadataError
-from django_modern_rest.serialization import BaseSerializer
+from django_modern_rest.serializer import BaseSerializer
 
 if TYPE_CHECKING:
     from django_modern_rest.controller import Blueprint, Controller
