+++ django_test_app/server/apps/controllers/auth.py
@@ -10,7 +10,7 @@ from django_modern_rest.security.http import (
 if TYPE_CHECKING:
     from django_modern_rest.controller import Controller
     from django_modern_rest.endpoint import Endpoint
-    from django_modern_rest.serialization import BaseSerializer
+    from django_modern_rest.serializer import BaseSerializer
 
 
 @final
