+++ docs/examples/error_handling/wrap_endpoint.py
@@ -7,7 +7,7 @@ from django_modern_rest import Body, Controller, modify
 from django_modern_rest.endpoint import Endpoint
 from django_modern_rest.errors import wrap_handler
 from django_modern_rest.plugins.pydantic import PydanticSerializer
-from django_modern_rest.serialization import BaseSerializer
+from django_modern_rest.serializer import BaseSerializer
 
 
 class TwoNumbers(pydantic.BaseModel):
