+++ docs/examples/using_controller/simple_routing_controllers.py
@@ -1,5 +1,5 @@
 from django_modern_rest.controller import Controller
-from django_modern_rest.serialization import BaseSerializer
+from django_modern_rest.serializer import BaseSerializer
 
 
 class UserList(Controller[BaseSerializer]):
