+++ tests/test_unit/test_controllers/test_customization/test_contoller_validator_customization.py
@@ -4,7 +4,7 @@ from typing_extensions import override
 
 from django_modern_rest import Blueprint, Controller
 from django_modern_rest.plugins.pydantic import PydanticSerializer
-from django_modern_rest.serialization import BaseSerializer
+from django_modern_rest.serializer import BaseSerializer
 from django_modern_rest.validation import BlueprintValidator
