+++ tests/test_unit/test_controllers/test_customization/test_endpoint_validator_customization.py
@@ -7,7 +7,7 @@ from typing_extensions import override
 from django_modern_rest import APIError, Blueprint, Controller
 from django_modern_rest.endpoint import Endpoint
 from django_modern_rest.plugins.pydantic import PydanticSerializer
-from django_modern_rest.serialization import BaseSerializer
+from django_modern_rest.serializer import BaseSerializer
 from django_modern_rest.test import DMRRequestFactory
 from django_modern_rest.validation import (
     EndpointMetadataValidator,
