+++ tests/test_unit/test_controllers/test_customization/test_serializer_context_customization.py
@@ -2,7 +2,7 @@ from typing import ClassVar, final
 
 from django_modern_rest import Controller
 from django_modern_rest.plugins.pydantic import PydanticSerializer
-from django_modern_rest.serialization import SerializerContext
+from django_modern_rest.serializer import SerializerContext
 
 
 @final
