+++ tests/test_unit/test_response/test_api_error.py
@@ -23,7 +23,7 @@ from django_modern_rest.openapi.objects.security_requirement import (
 )
 from django_modern_rest.plugins.pydantic import PydanticSerializer
 from django_modern_rest.security import AsyncAuth, SyncAuth
-from django_modern_rest.serialization import BaseSerializer
+from django_modern_rest.serializer import BaseSerializer
 from django_modern_rest.test import DMRAsyncRequestFactory, DMRRequestFactory
